Move NVI helper function
