+++ src/types/nvi.types.ts
@@ -89,8 +89,10 @@ export interface RejectedApproval extends FinalizedApproval {
 
 type NviAllowedOperation = 'approval/reject-candidate' | 'approval/approve-candidate' | 'approval/reset-approval';
 
+export type NviCandidateProblemType = 'UnverifiedCreatorExists' | 'UnverifiedCreatorFromOrganizationProblem';
+
 interface NviCandidateProblem {
-  type: 'UnverifiedCreatorExists' | 'UnverifiedCreatorFromOrganizationProblem';
+  type: NviCandidateProblemType;
 }
 
 export interface NviCandidate {
