+++ src/utils/nviHelpers.ts
@@ -1,5 +1,6 @@
 import { fetchOrganization } from '../api/cristinApi';
 import { Contributor } from '../types/contributor.types';
+import { NviCandidate } from '../types/nvi.types';
 import { BookRegistration } from '../types/publication_types/bookRegistration.types';
 import { ChapterRegistration } from '../types/publication_types/chapterRegistration.types';
 import { JournalRegistration } from '../types/publication_types/journalRegistration.types';
@@ -133,3 +134,10 @@ export const willResetNviStatuses = async (persistedRegistration: Registration, 
     return true;
   }
 };
+
+export const hasUnidentifiedContributorProblem = (nviCandidate: NviCandidate) =>
+  nviCandidate.problems &&
+  nviCandidate.problems.some(
+    (problem) =>
+      problem.type === 'UnverifiedCreatorExists' || problem.type === 'UnverifiedCreatorFromOrganizationProblem'
+  );
